Fixes needed after rollback of time=os_period_t.

+++ cs/src/host_packet.cpp
@@ -162,8 +162,8 @@ void PacketQueue::ProcessPacket(PacketBase* pkt) {
         delete pkt;
         return;
       }
-      LOG(INFO, "usb packet len %d, cmd %02x, data %02x, %02x, %02x ...",
-          in_pkt.size(), in_pkt[0], in_pkt.buf()[1], in_pkt.buf()[2], in_pkt.buf()[3]);
+      LOG(INFO, "usb packet len %lu, cmd %02x, data %02x, %02x, %02x ...",
+          (unsigned long)in_pkt.size(), in_pkt[0], in_pkt.buf()[1], in_pkt.buf()[2], in_pkt.buf()[3]);
     }
     switch (in_pkt[0]) {
     case CMD_PING: {
